requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bidict==0.21.2
Box2D==2.3.10
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
discord==1.0.1
discord.py==1.6.0
fasteners==0.16.3
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-WTF==0.15.1
glfw==2.3.0
graphviz==0.17
-e git+https://github.com/openai/gym@45902be4d082e4f2c289a449a30efbf76a0616a2#egg=gym
idna==2.10
imageio==2.9.0
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keyboard==0.13.5
kiwisolver==1.3.1
KuCoin==0.1
llvmlite==0.35.0
lockfile==0.12.2
MarkupSafe==2.0.1
matplotlib==3.3.4
mpl-finance==0.10.1
mplfinance==0.12.7a5
mujoco-py==2.0.2.8
multidict==5.1.0
# Editable install with no version control (neat-python==0.92)
-e d:\projects\python\neat\neat-python-0.92
numba==0.52.0
numpy==1.20.1
opencv-python==4.5.3.56
pandas==1.2.2
Pillow==8.1.0
ppaquette-gym-super-mario==0.0.8
pycparser==2.20
pyglet==1.5.21
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==4.2.1
python-kucoin==2.1.2
python-socketio==5.4.0
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
sklearn==0.0
soupsieve==2.2
threadpoolctl==2.2.0
torch==1.7.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
vk-api==11.9.1
websockets==8.1
Werkzeug==2.0.1
WTForms==2.3.3
yarl==1.6.3
zipp==3.6.0
